pyproject.toml
[tool.poetry]
name = "oriax"
version = "0.1.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
colorama = "^0.4.6"
dnspython = "^2.6.1"
python-whois = "^0.9.4"
shodan = "^1.31.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
bs4 = "^0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
